+++ src/gravity/gravity.cpp
@@ -42,9 +42,6 @@ Gravity::Gravity(MeshBlock *pmb, ParameterInput *pin)
   if (pmb->block_size.nx3 > 1) ncells3 = pmb->block_size.nx3 + 2*(NGHOST);
 
   phi.NewAthenaArray(ncells3,ncells2,ncells1);
-  if(SELF_GRAVITY_ENABLED == 1){
-    phi_old.NewAthenaArray(ncells3,ncells2,ncells1);
-  }
 }
 
 // destructor
@@ -52,8 +49,5 @@ Gravity::Gravity(MeshBlock *pmb, ParameterInput *pin)
 Gravity::~Gravity()
 {
   phi.DeleteAthenaArray();
-  if(SELF_GRAVITY_ENABLED == 1){
-    phi_old.DeleteAthenaArray();
-  }
 }
