+++ src/gravity/mggravity.cpp
@@ -8,12 +8,126 @@
 
 // Athena++ headers
 #include "mggravity.hpp"
+#include "gravity.hpp"
 #include "../athena.hpp"
 #include "../athena_arrays.hpp"
 #include "../mesh/mesh.hpp"
+#include "../hydro/hydro.hpp"
 #include "../coordinates/coordinates.hpp"
+#include "../parameter_input.hpp"
 #include "../multigrid/multigrid.hpp"
 #include "../globals.hpp"
+
+#include <iostream>
+#include <sstream>    // sstream
+#include <stdexcept>  // runtime_error
+#include <string>     // c_str()
+
+#ifdef MPI_PARALLEL
+#include <mpi.h>
+#endif
+
+class MeshBlock;
+
+//----------------------------------------------------------------------------------------
+//! \fn MGGravityDriver::MGGravityDriver(Mesh *pm, MGBoundaryFunc_t *MGBoundary,
+//                                   ParameterInput *pin)
+//  \brief MGGravityDriver constructor
+
+MGGravityDriver::MGGravityDriver(Mesh *pm, MGBoundaryFunc_t *MGBoundary, ParameterInput *pin)
+ : MultigridDriver(pm, MGBoundary, 1)
+{
+  four_pi_G_=pmy_mesh_->four_pi_G_;
+  eps_=pmy_mesh_->grav_eps_;
+  if(four_pi_G_==0.0) {
+   std::stringstream msg;
+   msg << "### FATAL ERROR in MGGravityDriver::MGGravityDriver" << std::endl
+        << "Gravitational constant must be set in the Mesh::InitUserMeshData "
+        << "using the SetGravitationalConstant or SetFourPiG function." << std::endl;
+    throw std::runtime_error(msg.str().c_str());
+    return;
+  }
+  if(mode_>=2 && eps_<0.0) {
+   std::stringstream msg;
+   msg << "### FATAL ERROR in MGGravityDriver::MGGravityDriver" << std::endl
+        << "Convergence threshold must be set in the Mesh::InitUserMeshData "
+        << "using the SetGravitatyThreshold for the iterative mode." << std::endl
+        << "Set the threshold = 0.0 for automatic convergence control." << std::endl;
+    throw std::runtime_error(msg.str().c_str());
+    return;
+  }
+
+  // Allocate multigrid objects
+  RegionSize root_size=pmy_mesh_->mesh_size;
+  root_size.nx1=pmy_mesh_->nrbx1;
+  root_size.nx2=pmy_mesh_->nrbx2;
+  root_size.nx3=pmy_mesh_->nrbx3;
+  LogicalLocation lroot;
+  lroot.lx1=0, lroot.lx2=0, lroot.lx3=0, lroot.level=0;
+  mgroot_= new MGGravity(this,lroot,-1,-1,root_size,MGBoundary,pmy_mesh_->mesh_bcs,true);
+  pmg_=NULL;
+  Multigrid *pfirst;
+  int nbs=nslist_[Globals::my_rank];
+  int nbe=nbs+nblist_[Globals::my_rank]-1;
+  RegionSize block_size;
+  block_size.nx1=pmy_mesh_->mesh_size.nx1/pmy_mesh_->nrbx1;
+  block_size.nx2=pmy_mesh_->mesh_size.nx2/pmy_mesh_->nrbx2;
+  block_size.nx3=pmy_mesh_->mesh_size.nx3/pmy_mesh_->nrbx3;
+  for(int i=nbs;i<=nbe;i++) {
+    enum BoundaryFlag block_bcs[6];
+    pmy_mesh_->SetBlockSizeAndBoundaries(pmy_mesh_->loclist[i], block_size, block_bcs);
+    Multigrid *nmg=new MGGravity(this, pmy_mesh_->loclist[i], i, i-nbs, block_size,
+                                 MGBoundary, block_bcs, false);
+    nmg->pmgbval->SearchAndSetNeighbors(pmy_mesh_->tree, ranklist_, nslist_);
+    AddMultigrid(nmg);
+  }
+}
+
+
+//----------------------------------------------------------------------------------------
+//! \fn void MGGravityDriver::Solve(int step)
+//  \brief load the data and solve
+
+void MGGravityDriver::Solve(int step)
+{
+  Multigrid *pmggrav=pmg_;
+  AthenaArray<Real> in;
+
+  // Load the source 
+  while(pmggrav!=NULL) {
+    MeshBlock *pmb=pmy_mesh_->FindMeshBlock(pmggrav->gid_);
+    if(pmb!=NULL) {
+      if(step==1) in.InitWithShallowCopy(pmb->phydro->u);
+      else if(step==2) in.InitWithShallowCopy(pmb->phydro->u1);
+      pmggrav->LoadSource(in, IDN, NGHOST, four_pi_G_);
+      if(mode_>=2) // iterative mode - load initial guess
+        pmggrav->LoadFinestData(pmb->pgrav->phi, 0, NGHOST);
+    }
+//    else { // on another process
+//    }
+    pmggrav=pmggrav->next;
+  }
+
+  SetupMultigrid();
+  Real mean_rho=0.0;
+  if(fperiodic_) mean_rho=last_ave_/four_pi_G_;
+  if(mode_<=1) SolveFMGCycle();
+  else SolveIterative();
+
+  // Return the result
+  pmggrav=pmg_;
+  while(pmggrav!=NULL) {
+    MeshBlock *pmb=pmy_mesh_->FindMeshBlock(pmggrav->gid_);
+    if(pmb!=NULL) {
+      pmggrav->RetrieveResult(pmb->pgrav->phi,0,2);
+      pmb->pgrav->grav_mean_rho=mean_rho;
+    }
+//    else { // on another process
+//    }
+    pmggrav=pmggrav->next;
+  }
+  return;
+}
 
 //----------------------------------------------------------------------------------------
 //! \fn  void MGGravity::Smooth(int color)
